Organizing and consolidating files in the repository
Created an INPUT and OUTPUT file folder. Renamed the TEST folder to UNIT TEST.
